chore: refactor serverless e2e tests
