--- e2e/helpers/cloud-run-telemetry-checker.ts
@@ -1,5 +1,0 @@
-import {checkTelemetryFlowing as checkE2ETelemetryFlowing} from './telemetry-checker'
-
-export const checkTelemetryFlowing = async (serviceName: string): Promise<void> => {
-  await checkE2ETelemetryFlowing({serviceName})
-}
