+++ e2e/serverless/aas-linux.test.ts
@@ -1,7 +1,8 @@
 import crypto from 'node:crypto'
 
+import {DATADOG_CI_COMMAND, execPromise, execPromiseWithRetries} from '../helpers/exec'
+
 import {verifyLinuxInstrumented, verifyLinuxUninstrumented} from './helpers/aas-verifier'
-import {DATADOG_CI_COMMAND, execPromise, execPromiseWithRetries} from './helpers/exec'
 import {checkTelemetryFlowing} from './helpers/telemetry-checker'
 import {triggerTraffic} from './helpers/traffic'
