+++ e2e/serverless/helpers/container-app-verifier.ts
@@ -1,4 +1,4 @@
-import {execSync} from './exec'
+import {execSync} from '../../helpers/exec'
 
 interface ContainerApp {
   properties: {
