+++ e2e/serverless/helpers/lambda-verifier.ts
@@ -1,4 +1,4 @@
-import {execSync} from './exec'
+import {execSync} from '../../helpers/exec'
 
 interface LambdaFunctionConfiguration {
   FunctionArn: string
