+++ e2e/serverless/helpers/telemetry-checker.ts
